[Silver IV] Title: 피하자, Time: 408 ms, Memory: 91492 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/25379. 피하자/피하자.cs
@@ -4,10 +4,10 @@
 int n = int.Parse(Console.ReadLine());
 int[] inputArr = Array.ConvertAll(Console.ReadLine().Split(), int.Parse);
 
-int leftSort = 0;
-int rightSort = 0;
-int leftCount = 0;
-int rightCount = 0;
+long leftSort = 0;
+long rightSort = 0;
+long leftCount = 0;
+long rightCount = 0;
 for (int i = 0; i < n; i++)
 {
     if (inputArr[i] % 2 == 0)
